still a transitional state, but ready for a preview release
NOTE: job_description style is not working, this will be the example style for the preprocessor stage

new exporter "Rich Export". it exports a file not unlike the archimate model file, but uses xsi;type (the element type) as tag name
I plan further development on this

some restructuring of code

the styled export now uses rich export format as a starting point. the provided stylesheets are rewritten according to this.

the styled export now can have a preprocessor, either in xlst or python. when I will have a fully working example of it, I will document the feature.

the export functionalities now try to remember the last saved file name

README for the default style

new style: sketchtotable. intended to be facilitate the use of Archi in Nominal Group technique