policy support, documentation, postprocessor
This is a preview release. Things work, but not everything is in right place.
Error console is missing, and the documentation is lacking.

We have (some) documentation now. The documentation is an archimate model, of course.

There is policy support. It is based on objectClasses. See the part about objectClasses for details.
The errors and warnings generated by policy check go to stdout now. I plan to change this, but I don't yet have a clue on how to implement it:)
The warnings are not detailed (until there will be correct error console).

Now we have postprocessing support. You can use either an xslt or a python script (postprocess.{xslt|py}).

The documentcontrol style now includes a documentation on objectClasses, and also generates the policy from the model.
This generated policy is used for the style.

The documentcontrol style contains a simple python postprocessor, which cuts the policy out from the generated report and save it away. It can serve an example on how to write a pre/postprocessor in python.