restructure of reporting
This is a work in progress release (as usual). I won't have any time to play with it for two weeks.
Now a .style file drives the reporting: you can generate more output files in one run,
and can even do tricks like first generate the policy, and then validate the model against it.
Trying to make the provided xslts more reuseable.